8b16774 Fixing Visual Studio compiler warning
14559ba Fixing return values in vtkDataObjectWriter
88cb75d Include more API from vtkDataWriter in vtkDataObjectWriter

+++ IO/vtkDataReader.h
@@ -80,7 +80,7 @@ public:
   vtkGetMacro(InputStringLength, int);
   void SetBinaryInputString(const char *, int len);
   void SetInputString(const vtkStdString& input)
-    { this->SetBinaryInputString(input.c_str(), input.length()); }
+    { this->SetBinaryInputString(input.c_str(), static_cast<int>(input.length())); }
 
   // Description:
   // Specify the vtkCharArray to be used  when reading from a string.
